qMiner/requirements.txt
Flask==3.0.2
aiohttp==3.10.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
aiosqlite==0.20.0
asgiref==3.8.1